Python 3.10 fix

+++ blink/aboutpanel.py
@@ -50,9 +50,9 @@ class AboutPanel(base_class, ui_class, metaclass=QSingleton):
 
     def _credits_size_changed(self, size):
         self.credits_text.document().documentLayout().documentSizeChanged.disconnect(self._credits_size_changed)
-        self.setFixedSize(self.minimumSize().width(), self.minimumSize().width()*1.40)  # set a fixed aspect ratio
+        self.setFixedSize(self.minimumSize().width(), int(self.minimumSize().width() * 1.40))  # set a fixed aspect ratio
         row_height = self.credits_text.fontMetrics().height() + 2  # +2 for cellspacing
-        max_credits_height = 8*row_height + 2 + 14  # allow for maximum 8 rows; +2 for cellspacing and +14 for top/bottom margins
+        max_credits_height = 8 * row_height + 2 + 14  # allow for maximum 8 rows; +2 for cellspacing and +14 for top/bottom margins
         if self.credits_text.height() > max_credits_height:
             self.setFixedHeight(self.height() - (self.credits_text.height() - max_credits_height))
